Skeleton analyzer recognizes skeletons with invalid crossroads.
Previous version expected that in a skeleton each white pixel
has at most three neighbours and that neighbours and that there
is no cycle. But in some cases there can be a cycle:
X
 X
 XX
X  X
Note that the three pixels in the middle are all crossroad pixels
that form a cycle. Unfortunately they should be counted as one
crossroad. In the new implementation only one of them will be
marked as crossroad. Which one depends on direction from which
the algorithm enters.